+++ src/utils.ts
@@ -1,3 +1,6 @@
+import { basename } from 'path';
+import { fromFile } from 'hasha';
+
 export function determine_api_base_url(sandbox: boolean) {
   if (sandbox) {
     return 'https://sandbox.zenodo.org/api';
@@ -9,4 +12,41 @@ export function auth_headers(access_token: string) {
   return {
     Authorization: `Bearer ${access_token}`,
   };
+}
+
+export interface DepositionFile {
+  id: string;
+  filename: string;
+  filesize: number;
+  checksum: string;
+}
+
+export class FilePresentError extends Error {
+  constructor(m: string) {
+    super(m);
+    Object.setPrototypeOf(this, FilePresentError.prototype);
+  }
+}
+
+export class UnspportedHashingAlgorithm extends Error {
+  constructor(m: string) {
+    super(m);
+    Object.setPrototypeOf(this, UnspportedHashingAlgorithm.prototype);
+  }
+}
+
+export async function is_same_file_present(
+  file: string,
+  bag: DepositionFile[]
+) {
+  const filename = basename(file);
+  const other = bag.find(f => f.filename === filename);
+  if (!other) {
+    return false;
+  }
+  const checksum = await fromFile(file, { algorithm: 'md5' });
+  if (checksum === other.checksum) {
+    throw new FilePresentError(`File ${file} is present`);
+  }
+  return true;
 }